feat: add Clarke-Wright Algorithm in OperationController

+++ Backend-RouterCraft/database/migrations/2024_06_24_005514_create_clients_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable();
             // $table->string('email')->unique();
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->float('latitude');
+            $table->float('longitude');
             $table->timestamps();
             $table->unsignedBigInteger('operation_id');
             $table->foreign('operation_id')->references('id')->on('operations')->onDelete('cascade');
